Add failing tests
