+++ test/test_index.py
@@ -37,14 +37,7 @@ from git.index.util import TemporaryFileSwap
 from git.objects import Blob
 from git.util import Actor, cwd, hex_to_bin, rmtree
 
-from test.lib import (
-    TestBase,
-    VirtualEnvironment,
-    fixture,
-    fixture_path,
-    with_rw_directory,
-    with_rw_repo,
-)
+from test.lib import TestBase, VirtualEnvironment, fixture, fixture_path, with_rw_directory, with_rw_repo, PathLikeMock
 
 HOOKS_SHEBANG = "#!/usr/bin/env sh\n"
 
@@ -587,11 +580,15 @@ class TestIndex(TestBase):
                     yield entry.path
                 elif type_id == 1:  # path (PathLike)
                     yield Path(entry.path)
-                elif type_id == 2:  # blob
+                elif type_id == 2:  # path mock (PathLike)
+                    yield PathLikeMock(entry.path)
+                elif type_id == 3:  # path mock in a blob
                     yield Blob(rw_repo, entry.binsha, entry.mode, entry.path)
-                elif type_id == 3:  # BaseIndexEntry
+                elif type_id == 4:  # blob
+                    yield Blob(rw_repo, entry.binsha, entry.mode, entry.path)
+                elif type_id == 5:  # BaseIndexEntry
                     yield BaseIndexEntry(entry[:4])
-                elif type_id == 4:  # IndexEntry
+                elif type_id == 6:  # IndexEntry
                     yield entry
                 else:
                     raise AssertionError("Invalid Type")
@@ -1198,13 +1195,31 @@ class TestIndex(TestBase):
             raise AssertionError("Should have caught a HookExecutionError")
 
     @with_rw_repo("HEAD")
-    def test_index_add_pathlike(self, rw_repo):
+    def test_index_add_pathlib(self, rw_repo):
         git_dir = Path(rw_repo.git_dir)
 
         file = git_dir / "file.txt"
         file.touch()
 
         rw_repo.index.add(file)
+
+    @with_rw_repo("HEAD")
+    def test_index_add_pathlike(self, rw_repo):
+        git_dir = Path(rw_repo.git_dir)
+
+        file = git_dir / "file.txt"
+        file.touch()
+
+        rw_repo.index.add(PathLikeMock(str(file)))
+
+    @with_rw_repo("HEAD")
+    def test_index_add_pathlike_unicode(self, rw_repo):
+        git_dir = Path(rw_repo.git_dir)
+
+        file = git_dir / "file-áēñöưḩ̣.txt"
+        file.touch()
+
+        rw_repo.index.add(PathLikeMock(str(file)))
 
     @with_rw_repo("HEAD")
     def test_index_add_non_normalized_path(self, rw_repo):
